Extract common base class to avoid test duplication
